+++ example/android/app/src/main/kotlin/com/example/flutter_speed_dial_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.flutter_speed_dial_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}