+++ lib/src/speed_dial.dart
@@ -1,11 +1,15 @@
+import 'dart:math';
+
 import 'package:flutter/material.dart';
 
 import 'animated_child.dart';
 import 'animated_floating_button.dart';
 import 'background_overlay.dart';
 import 'speed_dial_child.dart';
+import 'speed_dial_orientation.dart';
 
 /// Builds the Speed Dial
+// ignore: must_be_immutable
 class SpeedDial extends StatefulWidget {
   /// Children buttons, from the lowest to the highest.
   final List<SpeedDialChild> children;
@@ -21,10 +25,11 @@ class SpeedDial extends StatefulWidget {
   final Color backgroundColor;
   final Color foregroundColor;
   final double elevation;
+  final double buttonSize;
   final ShapeBorder shape;
   final Gradient gradient;
 
-  final double marginRight;
+  final double marginEnd;
   final double marginBottom;
 
   /// The color of the background overlay.
@@ -39,8 +44,26 @@ class SpeedDial extends StatefulWidget {
   /// The theme for the animated icon.
   final IconThemeData animatedIconTheme;
 
-  /// The child of the main button, ignored if [animatedIcon] is non [null].
-  final Widget child;
+  /// The icon of the main button, ignored if [animatedIcon] is non [null].
+  final IconData icon;
+
+  /// The active icon of the main button, Defaults to icon if not specified, ignored if [animatedIcon] is non [null].
+  final IconData activeIcon;
+
+  /// If true then rotation animation will be used when animating b/w activeIcon and icon.
+  final bool useRotationAnimation;
+
+  /// The theme for the icon generally includes color and size.
+  final IconThemeData iconTheme;
+
+  /// The label of the main button.
+  final Widget label;
+
+  /// The active label of the main button, Defaults to label if not specified.
+  final Widget activeLabel;
+
+  /// Transition Builder between label and activeLabel, defaults to FadeTransition.
+  final Widget Function(Widget, Animation<double>) labelTransitionBuilder;
 
   /// Executed when the dial is opened.
   final VoidCallback onOpen;
@@ -54,39 +77,91 @@ class SpeedDial extends StatefulWidget {
   /// If true user is forced to close dial manually by tapping main button. WARNING: If true, overlay is not rendered.
   final bool closeManually;
 
-  /// The speed of the animation
+  /// If true overlay is rendered, no matter if closeManually is true or false.
+  final bool renderOverlay;
+
+  /// Open or close the dial via a notification
+  final ValueNotifier<bool> openCloseDial;
+
+  /// The speed of the animation in milliseconds
   final int animationSpeed;
 
-  SpeedDial(
-      {this.children = const [],
-      this.visible = true,
-      this.backgroundColor,
-      this.foregroundColor,
-      this.gradient,
-      this.elevation = 6.0,
-      this.overlayOpacity = 0.8,
-      this.overlayColor = Colors.white,
-      this.tooltip,
-      this.heroTag,
-      this.animatedIcon,
-      this.animatedIconTheme,
-      this.child,
-      this.marginBottom = 16,
-      this.marginRight = 16,
-      this.onOpen,
-      this.onClose,
-      this.closeManually = false,
-      this.shape = const CircleBorder(),
-      this.curve = Curves.linear,
-      this.onPress,
-      this.animationSpeed = 150});
+  /// The bottom margin of each child
+  final double childMarginBottom;
+
+  /// The top margin of each child
+  final double childMarginTop;
+
+  /// The orientation of the children. Default is [SpeedDialOrientation.Up]
+  final SpeedDialOrientation orientation;
+
+  /// If Provided then it will replace the default Floating Action Button
+  /// and will show the Widget Specified as dialRoot instead, it will also
+  /// ignore backgroundColor, foregroundColor or any other property
+  /// that was specific to FAB before like onPress, you will have to provide
+  /// it again to your dialRoot button.
+  final Widget dialRoot;
+
+  /// If Provided then it will use Inkwell
+  /// for onLongPress instead of GestureDetector on Top of Root Widget
+  final bool useInkWell;
+
+  /// This is the child of the FAB, if specified it will ignore icon, activeIcon.
+  final Widget child;
+
+  /// This is the active child of the FAB, if specified it will animate b/w this
+  /// and the child.
+  final Widget activeChild;
+
+  SpeedDial({
+    Key key,
+    this.children = const [],
+    this.visible = true,
+    this.backgroundColor,
+    this.foregroundColor,
+    this.gradient,
+    this.elevation = 6.0,
+    this.buttonSize = 56.0,
+    this.dialRoot,
+    this.useInkWell = false,
+    this.overlayOpacity = 0.8,
+    this.overlayColor,
+    this.tooltip,
+    this.heroTag,
+    this.animatedIcon,
+    this.animatedIconTheme,
+    this.icon,
+    this.activeIcon,
+    this.child,
+    this.activeChild,
+    this.useRotationAnimation = true,
+    this.iconTheme,
+    this.label,
+    this.activeLabel,
+    this.labelTransitionBuilder,
+    this.marginBottom = 16,
+    this.marginEnd = 16,
+    this.onOpen,
+    this.onClose,
+    this.orientation = SpeedDialOrientation.Up,
+    this.closeManually = false,
+    this.renderOverlay = false,
+    this.shape = const CircleBorder(),
+    this.curve = Curves.linear,
+    this.onPress,
+    this.animationSpeed = 150,
+    this.openCloseDial,
+    this.childMarginBottom = 0,
+    this.childMarginTop = 0,
+  }) : super(key: key);
 
   @override
   _SpeedDialState createState() => _SpeedDialState();
+
+  bool _dark;
 }
 
-class _SpeedDialState extends State<SpeedDial>
-    with SingleTickerProviderStateMixin {
+class _SpeedDialState extends State<SpeedDial> with TickerProviderStateMixin {
   AnimationController _controller;
 
   bool _open = false;
@@ -98,6 +173,12 @@ class _SpeedDialState extends State<SpeedDial>
       duration: _calculateMainControllerDuration(),
       vsync: this,
     );
+    widget.openCloseDial?.addListener(() {
+      final show = widget.openCloseDial?.value;
+      if (_open != show) {
+        _toggleChildren();
+      }
+    });
   }
 
   Duration _calculateMainControllerDuration() => Duration(
@@ -129,13 +210,16 @@ class _SpeedDialState extends State<SpeedDial>
   }
 
   void _toggleChildren() {
-    var newValue = !_open;
-    setState(() {
-      _open = newValue;
-    });
-    if (newValue && widget.onOpen != null) widget.onOpen();
-    _performAnimation();
-    if (!newValue && widget.onClose != null) widget.onClose();
+    if (widget.children.length > 0) {
+      var newValue = !_open;
+      setState(() {
+        _open = newValue;
+      });
+      if (widget.openCloseDial != null) widget.openCloseDial.value = newValue;
+      if (newValue && widget.onOpen != null) widget.onOpen();
+      _performAnimation();
+      if (!newValue && widget.onClose != null) widget.onClose();
+    } else if (widget.onOpen != null) widget.onOpen();
   }
 
   List<Widget> _getChildrenList() {
@@ -145,7 +229,8 @@ class _SpeedDialState extends State<SpeedDial>
         .map((SpeedDialChild child) {
           int index = widget.children.indexOf(child);
 
-          var childAnimation = Tween(begin: 0.0, end: 62.0).animate(
+          var childAnimation =
+              Tween(begin: 0.0, end: widget.buttonSize).animate(
             CurvedAnimation(
               parent: this._controller,
               curve: Interval(0, singleChildrenTween * (index + 1)),
@@ -155,16 +240,21 @@ class _SpeedDialState extends State<SpeedDial>
           return AnimatedChild(
             animation: childAnimation,
             index: index,
+            key: child.key,
             visible: _open,
+            useInkWell: widget.useInkWell,
+            dark: widget._dark,
             backgroundColor: child.backgroundColor,
             foregroundColor: child.foregroundColor,
             elevation: child.elevation,
+            buttonSize: widget.buttonSize,
             child: child.child,
             label: child.label,
             labelStyle: child.labelStyle,
             labelBackgroundColor: child.labelBackgroundColor,
             labelWidget: child.labelWidget,
             onTap: child.onTap,
+            onLongPress: child.onLongPress,
             toggleChildren: () {
               if (!widget.closeManually) _toggleChildren();
             },
@@ -172,6 +262,8 @@ class _SpeedDialState extends State<SpeedDial>
             heroTag: widget.heroTag != null
                 ? '${widget.heroTag}-child-$index'
                 : null,
+            childMarginBottom: widget.childMarginBottom,
+            childMarginTop: widget.childMarginTop,
           );
         })
         .toList()
@@ -180,16 +272,17 @@ class _SpeedDialState extends State<SpeedDial>
   }
 
   Widget _renderOverlay() {
-    return Positioned(
-      right: -16.0,
+    return PositionedDirectional(
+      end: -16.0,
       bottom: -16.0,
       top: _open ? 0.0 : null,
-      left: _open ? 0.0 : null,
+      start: _open ? 0.0 : null,
       child: GestureDetector(
         onTap: _toggleChildren,
         child: BackgroundOverlay(
           animation: _controller,
-          color: widget.overlayColor,
+          color: widget.overlayColor ??
+              (widget._dark ? Colors.grey[900] : Colors.white),
           opacity: widget.overlayOpacity,
         ),
       ),
@@ -210,60 +303,144 @@ class _SpeedDialState extends State<SpeedDial>
               ),
             ),
             decoration: BoxDecoration(
-              shape: BoxShape.circle,
+              shape: shape,
               gradient: widget.gradient,
             ),
-          )
-        : widget.child;
+        )
 
-    var fabChildren = _getChildrenList();
+        : AnimatedBuilder(
+            animation: _controller,
+            builder: (BuildContext context, Widget _widget) => Transform(
+              transform: Matrix4.rotationZ(_controller.value * 0.5 * pi),
+              alignment: FractionalOffset.center,
+              child: AnimatedSwitcher(
+                duration: Duration(milliseconds: widget.animationSpeed),
+                child: (widget.activeChild == null &&
+                        widget.child != null &&
+                        _controller.value < 0.5)
+                    ? widget.child
+                    : (widget.activeChild != null && _controller.value > 0.5)
+                        ? widget.activeChild
+                        : (widget.activeIcon == null || _controller.value < 0.5)
+                            ? Icon(
+                                widget.icon,
+                                key: ValueKey<int>(0),
+                                color: widget.iconTheme?.color,
+                                size: widget.iconTheme?.size,
+                              )
+                            : Icon(
+                                widget.activeIcon,
+                                key: ValueKey<int>(1),
+                                color: widget.iconTheme?.color,
+                                size: widget.iconTheme?.size,
+                              ),
+              ),
+            ),
+          );
+
+    var label = AnimatedSwitcher(
+      duration: Duration(milliseconds: widget.animationSpeed),
+      transitionBuilder: widget.labelTransitionBuilder != null
+          ? widget.labelTransitionBuilder
+          : (child, animation) => FadeTransition(
+                opacity: animation,
+                child: child,
+              ),
+      child: (!_open || widget.activeLabel == null)
+          ? widget.label
+          : widget.activeLabel,
+    );
+
+    var fabChildren = _open ? _getChildrenList() : [];
 
     var animatedFloatingButton = AnimatedFloatingButton(
+      key: widget.key,
       visible: widget.visible,
+      useInkWell: widget.useInkWell,
       tooltip: widget.tooltip,
-      backgroundColor: widget.backgroundColor,
-      foregroundColor: widget.foregroundColor,
+      dialRoot: widget.dialRoot,
+      backgroundColor: widget.backgroundColor ??
+          (widget._dark ? Colors.grey[800] : Colors.grey[50]),
+      foregroundColor: widget.foregroundColor ??
+          (widget._dark ? Colors.white : Colors.black),
       elevation: widget.elevation,
       onLongPress: _toggleChildren,
       callback:
           (_open || widget.onPress == null) ? _toggleChildren : widget.onPress,
+      size: widget.buttonSize,
+      label: widget.label != null ? label : null,
       child: child,
       heroTag: widget.heroTag,
       shape: widget.shape,
       curve: widget.curve,
     );
 
-    return Positioned(
-      bottom: widget.marginBottom - 16,
-      right: widget.marginRight - 16,
-      child: Container(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.end,
-          crossAxisAlignment: CrossAxisAlignment.end,
-          children: List.from(fabChildren)
-            ..add(
-              Container(
-                margin: EdgeInsets.only(top: 8.0, right: 2.0),
-                child: animatedFloatingButton,
-              ),
+    switch (widget.orientation) {
+      case SpeedDialOrientation.Down:
+        return PositionedDirectional(
+          top: MediaQuery.of(context).size.height -
+              56 -
+              (widget.marginBottom - 16),
+          end: widget.marginEnd - 16,
+          child: Container(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.end,
+              crossAxisAlignment: CrossAxisAlignment.end,
+              children: List.from(fabChildren.reversed)
+                ..insert(
+                    0,
+                    Container(
+                      margin: EdgeInsetsDirectional.only(bottom: 8.0, end: 2.0),
+                      child: animatedFloatingButton,
+                    )),
             ),
-        ),
-      ),
-    );
+          ),
+        );
+        break;
+      case SpeedDialOrientation.Up:
+      default:
+        return PositionedDirectional(
+          bottom: widget.marginBottom - 16,
+          end: widget.marginEnd - 16,
+          child: Container(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.end,
+              crossAxisAlignment: CrossAxisAlignment.end,
+              children: List.from(fabChildren)
+                ..add(Container(
+                  margin: EdgeInsetsDirectional.only(top: 8.0, end: 2.0),
+                  child: animatedFloatingButton,
+                )),
+            ),
+          ),
+        );
+        break;
+    }
   }
 
   @override
   Widget build(BuildContext context) {
+    widget._dark = MediaQuery.of(context).platformBrightness == Brightness.dark;
     final children = [
-      if (!widget.closeManually) _renderOverlay(),
+      if ((!widget.closeManually || widget.renderOverlay) &&
+          widget.children.length > 0)
+        _renderOverlay(),
       _renderButton(),
     ];
 
-    return Stack(
+    var stack = Stack(
       alignment: Alignment.bottomRight,
       fit: StackFit.expand,
-      overflow: Overflow.visible,
+      clipBehavior: Clip.none,
       children: children,
     );
+
+    return (_open)
+        ? stack
+        : Container(
+            width: 56,
+            height: 56,
+            child: stack,
+          );
   }
 }