+++ CHANGELOG.md
@@ -1,13 +1,78 @@
+## [2.2.0] - 14 Feb 2021
+
+- Add child and activeChild
+- Fixed animation b/w non-active and active child's
+- Add useInkWell Property if you want to use Inkwell instead of GestureDetector
+- Add dialRoot Property to let you specify the root instead of the standard FAB button
+- Fixed formatting
+
+## [2.1.1] - 07 Feb 2021
+
+- README fixes
+
+## [2.1.0] - 04 Feb 2021
+
+- Add onLongPress support to SpeedDialChild by @m0veax
+- Add key support for both Speed Dial and Speed Dial Child
+
+## [2.0.0+1] - 31 Jan 2021
+
+- Fix Rotation Animation
+- Improve visibility and error related to it.
+- Add default onTap property to LabelWidget by default by @wongzq.
+- Add renderOverlay property to render overlay no matter what.
+
+## [2.0.0] - 30 Jan 2021
+
+- Add Label and activeLabel paramaters to SpeedDial
+- Made icon animation work
+- Add iconTheme
+- Add RTL support by @jacklebbos
+- Behave like FAB if no childrens are their by @Ionys320
+- Add openCloseDial parameter to programmatically control open and close the dial by @shizambles
+- Add option to customize orientation of the child buttons by @hinterlandsupplyco
+- Add childMarginBottom and childMarginTop option by @emavgl
+
+## [1.6.0] - 29 Jan 2021
+
+- Updated example project to androidX
+- Support borders other than CircularBorder by @m0veax
+- buttonSize parameter which defaults to 56 by @dwach414
+- Restrict Stack size to button size when closed by @tobilarscheid
+
+## [1.5.0] - 28 Jan 2021
+
+- icon and activeIcon instead of child
+- iconTransitionBuilder for custom transitions b/w icons
+- Add Dark Mode Support(@esieben1310) + Fix some issues related to it 
+- Centered child by @raviganwal
+- Only Build Speed Dial if childrens are visible by @davidmartos96
+
+## [1.3.0] - 16 Sept 2020
+
+- Used clipBehavior instead of Overflow as This feature was deprecated after v1.22.0-12.0.pre by @irasekh3
 
 ## [1.2.5] - 07 Dec 2019
 
+- Control Speed of Animations
+
 ## [1.2.4] - 28 Oct 2019
+
+- Fix multiple heroes with same 
+- Make label non required
 
 ## [1.2.3] - 20 Oct 2019
 
+- Fix Label Widget
+
 ## [1.2.2] - 18 Oct 2019
 
+- Add Label Widget
+
 ## [1.2.1] - 05 June 2019
+
+- Huge Optimization by removing unnecessary Animation Controllers
+- Other fixes
 
 ## [1.2.0] - 28 May 2019
 
@@ -15,4 +80,4 @@
 
 ## [0.0.1] - 07 July 2018
 
-* Initial Release
+- Initial Release